refactor(#342): 클래스 기간 수정 로직 리팩토링 및 폴더 위치 변경
